Towards modularizing ATen cmake (new sublibraries THRTS and THCRTS).
Right now, ATen is a big, monolithic cmake build system, which means that if
there are small pieces we want to reuse in other places, it is difficult to use
them (you essentially have to write a CMakeLists.txt for the relevant component
from scratch.)

This commit takes a step towards modularizing the cmake.  Specifically, it
takes the following steps:

- It bumps our minimum required CUDA version to 8.0, and eliminates some
  BC code for CUDA 8.0

- It extracts the common compile flag / feature testing preamble in
  aten/CMakeLists.txt and aten/src/ATen/CMakeLists.txt into a common
  header aten/cmake/Common.cmake.  This makes it easier to reuse this
  header in case we want other entry points to the build system.  Some
  of it is still a bit dirty; I only fixed things when they needed fixing.

- It extracts dependency finding and loading into aten/cmake/LoadFoo.cmake
  files.  This is not the desired end-state (the modern Caffe2 style is
  to make a new-style cmake target for each of our deps which we can
  depend on directly) but it's a good start.

- It splits out avx_mathfun.h and Intrinsics.h into a new package in src
  avx_mathfun, because these headers are depended upon by both TH and
  ATen directly.  I think this might be a temporary state of affairs
  because we will move things out of TH eventually.

- It turns TH into a static library, which will be linked into ATen.  This
  library declares its headers as public interfaces, so they are automatically
  picked up by libraries which depend on it, e.g., ATen proper (and more
  specifically the THC source.)

- It splits out a few files from TH into a new library THRTS.  The idea behind
  THRTS is that it is Torch's runtime system code: it contains key pieces such
  as the memory allocator that Torch ops need to run.  The actual operator code
  continues to live in TH.  This means that TH/THAtomic.h header includes no
  longer work; just say THAtomic.h instead.

- Similarly, it splits out a few files from THC into a new library THCRTS, with
  the same motivation.

- If one does the "obvious" cmake for linking static libraries, one quickly runs
  into the problem that linking a static library into a dynamic library causes
  any symbol which is not used to be stripped.  This is undesirable, because
  some symbols in our static library are used elsewhere.  Fortunately,
  caffe2 has a wonderful macro caffe2_interface_library which correctly
  handles the use of -Wl,--whole-archive to make this work correctly.

- It globally switches are target_link_libraries from old-style (keywordless)
  to new style (PUBLIC/PRIVATE/INTERFACE).  We did this because we need to mark
  any static libraries we link into libATen.so as PRIVATE, so that downstream uses
  of ATen don't attempt to also statically link them in.  This lead to a rather
  unusual error message where linking verify_api_visibility would lead to the
  message:

    /opt/conda/lib/libmkl_gf_lp64.so: undefined reference to `mkl_lapack_zgbtrf'

  Even though -lmkl_core was on the command line.  (No other executables would fail;
  verify_api_visibility failed because it linked against libATen.so but did not actually
  use any symbols.

  Basically, -llibATen.so -llibTH_static.a is WRONG WRONG WRONG and should never
  happen.  Making ATen's dependency on libTH PRIVATE solves the problem.

Testing strategy:

    cd aten/src/TH
    mkdir build
    cd build
    cmake ..
    make -j20

TH can be built by itself now.
