+++ aten/cmake/LoadARM.cmake
@@ -1,0 +1,17 @@
+# ARM specific flags
+find_package(ARM)
+if(ASIMD_FOUND)
+  message(STATUS "asimd/Neon found with compiler flag : -D__NEON__")
+  set(CMAKE_C_FLAGS "-D__NEON__ ${CMAKE_C_FLAGS}")
+elseif()
+  message(STATUS "Neon found with compiler flag : -mfpu=neon -D__NEON__")
+  set(CMAKE_C_FLAGS "-mfpu=neon -D__NEON__ ${CMAKE_C_FLAGS}")
+endif()
+if(CORTEXA8_FOUND)
+  message(STATUS "Cortex-A8 Found with compiler flag : -mcpu=cortex-a8")
+  set(CMAKE_C_FLAGS "-mcpu=cortex-a8 -fprefetch-loop-arrays ${CMAKE_C_FLAGS}")
+endif()
+if(CORTEXA9_FOUND)
+  message(STATUS "Cortex-A9 Found with compiler flag : -mcpu=cortex-a9")
+  set(CMAKE_C_FLAGS "-mcpu=cortex-a9 ${CMAKE_C_FLAGS}")
+endif()
