+++ aten/cmake/TestForCudaHalf.cmake
@@ -1,0 +1,7 @@
+if(CUDA_HAS_FP16 OR NOT ${CUDA_VERSION} LESS 7.5)
+  message(STATUS "Found CUDA with FP16 support, compiling with torch.CudaHalfTensor")
+  list(APPEND CUDA_NVCC_FLAGS "-DCUDA_HAS_FP16=1 -D__CUDA_NO_HALF_OPERATORS__")
+  set(CMAKE_C_FLAGS "-DCUDA_HAS_FP16=1 ${CMAKE_C_FLAGS}")
+else()
+  message(STATUS "Could not find CUDA with FP16 support, compiling without torch.CudaHalfTensor")
+endif()
