+++ aten/src/THCRTS/THCAllocator.h
@@ -1,7 +1,7 @@
 #ifndef THC_ALLOCATOR_INC
 #define THC_ALLOCATOR_INC
 
-#include "THCGeneral.h"
+#include "THCRTSGeneral.h"
 
 extern THAllocator THCudaHostAllocator;
 extern THAllocator THCUVAAllocator;
