+++ aten/src/THCRTS/THCRTSGeneral.cpp
@@ -1,0 +1,23 @@
+#include "THCRTSGeneral.h"
+
+void __THCudaCheck(cudaError_t err, const char *file, const int line)
+{
+  if(err != cudaSuccess)
+  {
+    static int alreadyFailed = 0;
+    if(!alreadyFailed) {
+      fprintf(stderr, "THCudaCheck FAIL file=%s line=%i error=%i : %s\n", file, line, err, cudaGetErrorString(err));
+      alreadyFailed = 1;
+    }
+    _THError(file, line, "cuda runtime error (%d) : %s", err,
+             cudaGetErrorString(err));
+  }
+}
+
+void __THCudaCheckWarn(cudaError_t err, const char *file, const int line)
+{
+  if(err != cudaSuccess)
+  {
+    fprintf(stderr, "THCudaCheckWarn FAIL file=%s line=%i error=%i : %s\n", file, line, err, cudaGetErrorString(err));
+  }
+}
