+++ aten/src/THCRTS/THCStream.cpp
