+++ aten/src/THRTS/THAtomic.c
