+++ setup.py
@@ -213,10 +213,10 @@ def build_libs(libs):
     if not IS_WINDOWS:
         if WITH_NINJA:
             my_env["CMAKE_GENERATOR"] = '-GNinja'
-            my_env["CMAKE_INSTALL"] = 'ninja install'
+            my_env["CMAKE_INSTALL"] = 'ninja install -v'
         else:
             my_env['CMAKE_GENERATOR'] = ''
-            my_env['CMAKE_INSTALL'] = 'make install'
+            my_env['CMAKE_INSTALL'] = 'make install VERBOSE=1'
     if WITH_SYSTEM_NCCL:
         my_env["NCCL_ROOT_DIR"] = NCCL_ROOT_DIR
     if WITH_CUDA:
